Refactoring binary load. load.py is now a subdir module.
