+++ mhdata/merge/binary/load/equipment_bload.py
@@ -1,35 +1,12 @@
 from typing import Type, Mapping, Iterable
-from os.path import dirname, abspath, join
-import re
 
-from mhdata import cfg
-from mhdata.util import OrderedSet, Sharpness, bidict
-
-from mhw_armor_edit import ftypes
-from mhw_armor_edit.ftypes import gmd, am_dat, arm_up, kire, wp_dat, wp_dat_g, eq_crt, eq_cus, skl_pt_dat
+from mhw_armor_edit.ftypes import gmd, am_dat, arm_up, kire, wp_dat, wp_dat_g, eq_crt, eq_cus
 from mhw_armor_edit.ftypes.ext import rod_inse
 
-# Location of MHW binary data.
-# Looks for a folder called /mergedchunks neighboring the main project folder.
-# This folder should be created via WorldChunkTool, with each numbered chunk being
-# moved into the mergedchunks folder in ascending order (with overwrite).
-CHUNK_DIRECTORY = join(dirname(abspath(__file__)), "../../../../mergedchunks")
+from mhdata import cfg
+from mhdata.util import Sharpness
 
-# Mapping from GMD filename suffix to actual language code
-lang_map = {
-    'eng': 'en',
-    'jpn': 'ja',
-    'fre': 'fr',
-    'ger': 'de',
-    'ita': 'it',
-    'spa': 'es',
-    'ptB': 'pt',
-    'pol': 'pl',
-    'rus': 'ru',
-    'kor': 'ko',
-    'chT': 'zh',
-    'ara': 'ar',
-}
+from .bcore import load_schema, load_text
 
 # wp_dat files (mapping from filename -> mhwdb weapon type)
 # ranged ones map to wp_dat_g instead
@@ -57,88 +34,6 @@ weapon_types = [
     cfg.HAMMER, cfg.HUNTING_HORN, cfg.LANCE, cfg.GUNLANCE, cfg.SWITCH_AXE,
     cfg.CHARGE_BLADE, cfg.INSECT_GLAIVE, cfg.BOW, cfg.HEAVY_BOWGUN, cfg.LIGHT_BOWGUN
 ]
-
-def load_schema(schema: Type[ftypes.StructFile], relative_dir: str) -> ftypes.StructFile:
-    "Uses an ftypes struct file class to load() a file relative to the chunk directory"
-    with open(join(CHUNK_DIRECTORY, relative_dir), 'rb') as f:
-        return schema.load(f)
-
-def load_text(basepath: str) -> Mapping[int, Mapping[str, str]]:
-    """Parses a series of GMD files, returning a mapping from index -> language -> value
-    
-    The given base path is the relative directory from the chunk folder,
-    excluding the _eng.gmd ending. All GMD files starting with the given basepath
-    and ending with the language are combined together into a single result.
-    """
-    results = {}
-    for ext_lang, lang in lang_map.items():
-        data = load_schema(gmd.Gmd, f"{basepath}_{ext_lang}.gmd")
-        for idx, value_obj in enumerate(data.items):
-            if idx not in results:
-                results[idx] = {}
-            value = value_obj.value
-            value = re.sub(r"( )*\r?\n( )*", " ", value)
-            value = re.sub(r"( )?<ICON ALPHA>", " α", value)
-            value = re.sub(r"( )?<ICON BETA>", " β", value)
-            value = re.sub(r"( )?<ICON GAMMA>", " γ", value)
-            results[idx][lang] = (value
-                                    .replace("<STYL MOJI_YELLOW_DEFAULT>[1]</STYL>", "[1]")
-                                    .replace("<STYL MOJI_YELLOW_DEFAULT>[2]</STYL>", "[2]")
-                                    .replace("<STYL MOJI_YELLOW_DEFAULT>[3]</STYL>", "[3]")
-                                    .replace("<STYL MOJI_YELLOW_DEFAULT>", "")
-                                    .replace("<STYL MOJI_LIGHTBLUE_DEFAULT>", "")
-                                    .replace("</STYL>", "")).strip()
-    return results
-
-class ItemTextHandler():
-    "A class that loads item text and tracks encountered items"
-
-    def __init__(self):
-        self._item_text = load_text("common/text/steam/item")
-        self.encountered = OrderedSet()
-
-    def name_for(self, item_id: int):
-        self.encountered.add(item_id)
-        return self._item_text[item_id * 2]
-
-    def description_for(self, item_id: int):
-        self.encountered.add(item_id)
-        return self._item_text[item_id * 2 + 1]
-
-    def text_for(self, item_id: int):
-        self.encountered.add(item_id)
-        return (self._item_text[item_id * 2], self._item_text[item_id * 2 + 1])
-
-def convert_recipe(item_text_handler: ItemTextHandler, recipe_binary) -> dict:
-    "Converts a recipe binary (of type eq_cus/eq_crt) to a dictionary"
-    new_data = {}
-    
-    for i in range(1, 4+1):
-        item_id = getattr(recipe_binary, f'item{i}_id')
-        item_qty = getattr(recipe_binary, f'item{i}_qty')
-
-        item_name = None if item_qty == 0 else item_text_handler.name_for(item_id)['en']
-        new_data[f'item{i}_name'] = item_name
-        new_data[f'item{i}_qty'] = item_qty if item_qty else None
-
-    return new_data
-
-class SkillTextHandler():
-    def __init__(self):    
-        self.skilltree_text = load_text("common/text/vfont/skill_pt")
-        
-        # mapping from name -> skill tree entry
-        self.skill_map = bidict()
-        for entry in load_schema(skl_pt_dat.SklPtDat, "common/equip/skill_point_data.skl_pt_dat").entries:
-            name = self.get_skilltree_name(entry.index)['en']
-            self.skill_map[name] = entry
-
-    def get_skilltree_name(self, skill_index: int) -> dict:
-        # Discovered formula via inspecting mhw_armor_edit's source.
-        return self.skilltree_text[skill_index * 3]
-
-    def get_skilltree(self, name_en: str) -> skl_pt_dat.SklPtDatEntry:
-        return self.skill_map[name_en]
 
 class SharpnessDataReader():
     "A class that loads sharpness data and processes it for binary weapon objects"
