+++ mhw_armor_edit/ftypes/ext/rod_inse.py
@@ -5,8 +5,8 @@ class RodInseEntry(Struct):
     STRUCT_SIZE = 28
     id: ft.uint()
     attack_type: ft.ubyte() # 0 = Server, 1 = Blunt
-    unk1: ft.ubyte()
-    unk2: ft.ubyte()
+    unk1: ft.ubyte() # unknown, could be GMD. Currently the same as id.
+    unk2: ft.ubyte() 
     base_model_id: ft.ushort()
     tree_id: ft.byte()
     cost: ft.uint()
@@ -15,9 +15,9 @@ class RodInseEntry(Struct):
     speed: ft.ushort()
     heal: ft.ushort()
     unk3: ft.ushort()
-    dust_type: ft.ubyte() # 0 = blast, 1 = heal, 2 = paralysis, 3 = poison
-    tree_position: ft.ushort()
-    unk4: ft.ushort()
+    dust_type: ft.ushort() # 0 = blast, 1 = heal, 2 = paralysis, 3 = poison
+    tree_position: ft.ubyte()
+    unk4: ft.ushort() # can't be GMD, as dragonsoul is OOB
 
 class RodInse(StructFile):
     "Kinsect file"
